Bug corrections and new features
The way some compilers treat expressions such as: long i = v.size() - 1, can be problematic when v.size() is 0, since in that case they will return not -1 but a much larger value due to way -1 is handled by unsigned long values.

We also add iferror, which replaces a try/catch.

+++ include/predicatefunctions.h
@@ -158,7 +158,7 @@ public:
 	}
 
 	bool Checkarity() {
-		return Arity(size, parameters.size() - 1);
+		return Arity(size, (long)parameters.size() - 1);
 	}
 
 };
