+++ src/tamgudvector.cxx
@@ -165,7 +165,7 @@ Exporting Tamgu* Tamgudvector::Pop(Tamgu* idx) {
         
     BLONG v = idx->Integer();
     if (v == -1) {
-        v = values.size() - 1;
+        v = (long)values.size() - 1;
         values.pop_back();
     }
     else {
